conflict Management
