+++ src/main/java/net/praqma/jenkins/plugin/prqa/globalconfig/QAVerifyServerConfiguration.java
@@ -9,12 +9,15 @@ import hudson.model.AbstractDescribableImpl;
 import hudson.model.Descriptor;
 import hudson.util.FormValidation;
 import net.praqma.jenkins.plugin.prqa.notifier.Messages;
+import org.apache.commons.lang.StringUtils;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.DataBoundSetter;
 import org.kohsuke.stapler.QueryParameter;
 
 import javax.annotation.Nonnull;
 import java.io.Serializable;
+import java.net.MalformedURLException;
+import java.net.URL;
 import java.util.Objects;
 
 
@@ -43,7 +46,8 @@ public class QAVerifyServerConfiguration
                                        String userName,
                                        String password,
                                        String protocol,
-                                       Integer viewerPortNumber) {
+                                       Integer viewerPortNumber,
+                                       String externalUrl) {
         this.configurationName = configurationName;
         this.hostName = hostName;
         this.password = password;
@@ -51,6 +55,7 @@ public class QAVerifyServerConfiguration
         this.portNumber = portNumber;
         this.protocol = protocol;
         this.viewerPortNumber = viewerPortNumber;
+        this.externalUrl = externalUrl;
     }
 
     /**
@@ -155,6 +160,13 @@ public class QAVerifyServerConfiguration
         return protocol + "://" + hostName + ":" + viewerPortNumber;
     }
 
+    public String getFullExternalUrl() {
+        if (StringUtils.isEmpty(externalUrl)) {
+            return getFullUrl();
+        }
+        return externalUrl;
+    }
+
     /**
      * @return the viewerPortNumber
      */
@@ -185,6 +197,15 @@ public class QAVerifyServerConfiguration
         this.protocol = protocol;
     }
 
+    public String getExternalUrl() {
+        return externalUrl;
+    }
+
+    @DataBoundSetter
+    public void setExternalUrl(String externalUrl) {
+        this.externalUrl = externalUrl;
+    }
+
     public enum ViewServerProtocol {
         http,
         https,
@@ -206,7 +227,7 @@ public class QAVerifyServerConfiguration
         @Nonnull
         @Override
         public String getDisplayName() {
-            return "QA·Verify Server";
+            return "QAÂ·Verify Server";
         }
 
         public FormValidation doCheckConfigurationName(@QueryParameter String value) {
@@ -227,6 +248,18 @@ public class QAVerifyServerConfiguration
 
         public FormValidation doCheckViewerPortNumber(@QueryParameter String value) {
             return checkValidPort(value);
+        }
+
+        public FormValidation doCheckExternalUrl(@QueryParameter String value) {
+            if (StringUtils.isEmpty(value)){
+                return FormValidation.ok();
+            }
+            try {
+                new URL(value);
+            } catch (MalformedURLException e) {
+                return FormValidation.error(e.getMessage());
+            }
+            return FormValidation.ok();
         }
 
         private FormValidation checkValidPort(@QueryParameter String value) {
